Port all commands to Sapphire
Finally using ~ as a prefix for all commands
Voice commands are still using primitive validation instead of
preconditions
