+++ src/index.ts
@@ -1,13 +1,9 @@
-import { SapphireClient } from '@sapphire/framework';
+import { AdobotClient } from './client/adobot_client.js';
 
-import { Intents, Message } from 'discord.js';
+
 import { loggers, transports } from 'winston';
 import { format } from 'logform';
 
-import { parseUserMessage } from './commands.js';
-import { VoiceInstanceDao } from './dao/voice_instance/base_dao.js';
-import { VoiceInstanceInMemoryDao } from './dao/voice_instance/in_memory_dao.js';
-import { AdobotClient } from './client/adobot_client.js';
 
 const DISCORD_TOKEN: string = process.env.ADOBOT_DISCORD_TOKEN !== undefined ? process.env.ADOBOT_DISCORD_TOKEN : '';
 const DEBUG_LEVEL: string = process.env.ADOBOT_DEBUG_LEVEL !== undefined ? process.env.ADOBOT_DEBUG_LEVEL : 'verbose';
@@ -37,14 +33,6 @@ client.on('ready', () => {
   client.user!.setActivity('adobo vids - yum!', { type: 'WATCHING' })
 });
 
-// parse messages
-client.on('messageCreate', async (msg: Message) => {
-  if (msg.author.id !== client.user!.id) {
-    // ignores messages sent by Adobot
-    await parseUserMessage(msg);
-  }
-
-});
 
 client.on('error', (err) => {
   logger.error(err);
